Refactor users app
