+++ posts/views.py
@@ -8,7 +8,6 @@ from django.utils import timezone
 from taggit.models import Tag
 
 from posts.models import Post, Space
-from users.forms import ProfileEditForm, UserEditForm
 from users.models import Contact
 
 from .forms import CommentForm, EmailPostForm, PostForm, SpaceCreationForm, SpaceForm
@@ -220,24 +219,6 @@ def my_research(request):
             "most_recent_posts": most_recent_posts,
             "most_commented_posts": most_commented_posts,
         },
-    )
-
-
-def my_account(request):
-    if request.method == "POST":
-        user_form = UserEditForm(instance=request.user, data=request.POST)
-        profile_form = ProfileEditForm(instance=request.user.profile, data=request.POST, files=request.FILES)
-        if user_form.is_valid() and profile_form.is_valid():
-            user_form.save()
-            profile_form.save()
-
-    else:
-        user_form = UserEditForm(instance=request.user)
-        profile_form = ProfileEditForm(instance=request.user.profile)
-    return render(
-        request,
-        "posts/my_account.html",
-        {"user_form": user_form, "profile_form": profile_form},
     )
