+++ users/urls.py
@@ -1,12 +1,21 @@
 from django.contrib.auth import views as auth_views
 from django.urls import path
 
-from . import views
+from users.views import (
+    edit,
+    login_request,
+    logout_request,
+    my_account,
+    register_request,
+    user_detail,
+    user_follow,
+    user_list,
+)
 
 urlpatterns = [
-    path("register/", views.register_request, name="register"),
-    path("login/", views.login_request, name="login"),
-    path("logout", views.logout_request, name="logout"),
+    path("register/", register_request, name="register"),
+    path("login/", login_request, name="login"),
+    path("logout", logout_request, name="logout"),
     path(
         "password_change/",
         auth_views.PasswordChangeView.as_view(template_name="password_change_form.html"),
@@ -32,8 +41,9 @@ urlpatterns = [
         "reset/done/",
         auth_views.PasswordResetCompleteView.as_view(template_name="password_reset_complete.html"),
     ),
-    path("edit/", views.edit, name="edit"),  # TODO: change name to more meaningful one
-    path("users/", views.user_list, name="user_list"),
-    path("users/<username>/", views.user_detail, name="user_detail"),
-    path("users/follow", views.user_follow, name="user_follow"),
+    path("edit/", edit, name="edit"),  # TODO: change name to more meaningful one
+    path("users/", user_list, name="user_list"),
+    path("users/<username>/", user_detail, name="user_detail"),
+    path("users/follow", user_follow, name="user_follow"),
+    path("myaccount/", my_account, name="my_account"),
 ]
