+++ users/views.py
@@ -148,3 +148,21 @@ def user_follow(request):
 def user_list(request):
     users = User.objects.filter(is_active=True)
     return render(request, "posts/user/list.html", {"section": "people", "users": users})
+
+
+def my_account(request):
+    if request.method == "POST":
+        user_form = UserEditForm(instance=request.user, data=request.POST)
+        profile_form = ProfileEditForm(instance=request.user.profile, data=request.POST, files=request.FILES)
+        if user_form.is_valid() and profile_form.is_valid():
+            user_form.save()
+            profile_form.save()
+
+    else:
+        user_form = UserEditForm(instance=request.user)
+        profile_form = ProfileEditForm(instance=request.user.profile)
+    return render(
+        request,
+        "posts/my_account.html",
+        {"user_form": user_form, "profile_form": profile_form},
+    )
